+++ src/main/java/dev/mardroide/fireworksmechanics/utils/EventActivation.java
@@ -1,0 +1,15 @@
+package dev.mardroide.fireworksmechanics.utils;
+
+import org.bukkit.Bukkit;
+
+public class EventActivation {
+    private static boolean isActivated;
+
+    public void setActivated(boolean activated) {
+        isActivated = activated;
+    }
+
+    public boolean isEventActivated() {
+        return isActivated;
+    }
+}