+++ src/main/java/dev/mardroide/fireworksmechanics/utils/GenerateRandomTrigger.java
@@ -3,7 +3,9 @@ package dev.mardroide.fireworksmechanics.utils;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class GenerateRandomTrigger {
-    public static double randomTrigger() {
-        return ThreadLocalRandom.current().nextDouble(0.00, 100.00);
+    double random = ThreadLocalRandom.current().nextDouble(0.00, 100.00);
+
+    public double getRandom() {
+        return random;
     }
 }