+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ### 1.0.0-alpha
 
-- Add early explode failure event
-- Add elytra boost fail event
+Fireworks:
+- Add early explode failure event.
+- Add elytra boost fail event.
+
+Dynamite:
+- Add primed entity explosion failure.
+- Add early explode failure event.